requirements.txt
boto3
psycopg2-binary
pytest
redis
requests
